Adding Normalmap support, ceiling normal and basic cobwebs.
The main environment shader now has normal map property and a basic normal. I've made a quick normal map for the ceiling, although it doesn't match perfectly.

I noticed the haunted house textures are not powers of 2 which will need to be fixed at some point.

Also I added a basic cartoony cobweb that blows in the wind. Wind speed and strength can be adjusted.